Relocate Sfinx spinner to shared components
