+++ app/shared/components/index.ts
@@ -5,3 +5,4 @@ export { default as Header } from "./Header";
 export { default as SfinxLogo } from "./SfinxLogo";
 export { default as UnauthGuard } from "./UnauthGuard";
 export { default as InterviewStageScreen } from "./InterviewStageScreen";
+export { default as SfinxSpinner } from "./SfinxSpinner";
